refactor: add OutletPortal
