+++ packages/mask/dashboard/pages/SetupPersona/Backup/Cloud/index.tsx
@@ -6,7 +6,7 @@ import { Box, Typography } from '@mui/material'
 import { memo } from 'react'
 import { Outlet, useMatch, useNavigate, useOutletContext } from 'react-router-dom'
 import { CloudBackupFormContext } from './CloudBackupFormContext.js'
-import type { PortalContainerProps } from '../../types.js'
+import type { PortalContainerProps } from '../../../../components/OutletPortal.js'
 
 const useStyles = makeStyles<void, 'activeButton'>()((theme, _, refs) => ({
     container: {
