+++ packages/mask/dashboard/pages/SetupPersona/Recovery/Cloud/GoogleDrive.tsx
@@ -3,12 +3,12 @@ import { Icons } from '@masknet/icons'
 import { EMPTY_LIST } from '@masknet/shared-base'
 import { makeStyles } from '@masknet/theme'
 import { GoogleDriveClient, type DriveFile } from '@masknet/web3-providers'
-import { Box, Button, Portal, Typography } from '@mui/material'
+import { Box, Button, Typography } from '@mui/material'
 import { memo, useMemo, useState } from 'react'
-import { useOutletContext } from 'react-router-dom'
 import { UserContext } from '../../../../../shared-ui/index.js'
 import { GoogleDriveFileTable } from '../../../../components/GoogleDriveFileTable.js'
 import { GoogleDriveLogin } from '../../../../components/GoogleDriveLogin.js'
+import { OutletPortal } from '../../../../components/OutletPortal.js'
 import { PrimaryButton } from '../../../../components/PrimaryButton/index.js'
 import { useGoogleDriveFiles } from '../../../../hooks/useGoogleDriveFiles.js'
 import { MergeBackupModal, RestoreBackupModal } from '../../../../modals/modals.js'
@@ -18,7 +18,6 @@ import {
     getGoogleDriveAccessToken,
     progressDownload,
 } from '../../../../utils/api.js'
-import type { PortalContainerProps } from '../../types.js'
 
 const useStyles = makeStyles()((theme) => ({
     container: {
@@ -64,7 +63,6 @@ const useStyles = makeStyles()((theme) => ({
 export const Component = memo(function GoogleDriveRecovery() {
     const { classes } = useStyles()
     const { user, updateUser } = UserContext.useContainer()
-    const { portalContainerRef } = useOutletContext<PortalContainerProps>()
     const googleDriveClient = useMemo(
         () => new GoogleDriveClient(getGoogleDriveAccessToken, clearGoogleDriveAccessToken),
         [],
@@ -129,7 +127,7 @@ export const Component = memo(function GoogleDriveRecovery() {
                     }}
                 />
             </Box>
-            <Portal container={() => portalContainerRef.current}>
+            <OutletPortal>
                 <PrimaryButton
                     variant="roundedContained"
                     startIcon={<Icons.Cloud size={18} />}
@@ -164,7 +162,7 @@ export const Component = memo(function GoogleDriveRecovery() {
                     }}>
                     <Trans>Recover</Trans>
                 </PrimaryButton>
-            </Portal>
+            </OutletPortal>
         </Box>
     )
 })
