+++ packages/mask/dashboard/pages/SetupPersona/types.ts
@@ -1,7 +1,3 @@
-export interface PortalContainerProps {
-    portalContainerRef: React.RefObject<HTMLDivElement | null>
-}
-
 export enum StorageType {
     Local = 'local',
     Cloud = 'cloud',
